Move sofa message logic into ChatManager
