+++ app/src/main/java/com/toshi/manager/SofaMessageManager.kt
@@ -18,17 +18,8 @@
 package com.toshi.manager
 
 
-import com.toshi.BuildConfig
-import com.toshi.R
 import com.toshi.crypto.HDWallet
-import com.toshi.crypto.signal.ChatService
-import com.toshi.crypto.signal.store.ProtocolStore
-import com.toshi.crypto.signal.store.SignalTrustStore
-import com.toshi.manager.chat.SofaMessageReceiver
-import com.toshi.manager.chat.SofaMessageRegistration
-import com.toshi.manager.chat.SofaMessageSender
-import com.toshi.manager.model.SofaMessageTask
-import com.toshi.manager.sofaMessageManager.SofaGroupManager
+import com.toshi.manager.chat.ChatManager
 import com.toshi.manager.store.ConversationStore
 import com.toshi.model.local.Conversation
 import com.toshi.model.local.ConversationObservables
@@ -36,199 +27,23 @@ import com.toshi.model.local.Group
 import com.toshi.model.local.IncomingMessage
 import com.toshi.model.local.Recipient
 import com.toshi.model.local.User
-import com.toshi.model.sofa.Init
-import com.toshi.model.sofa.SofaAdapters
 import com.toshi.model.sofa.SofaMessage
-import com.toshi.util.LocaleUtil
-import com.toshi.util.logging.LogUtil
-import com.toshi.util.sharedPrefs.SignalPrefs
 import com.toshi.view.BaseApplication
-import org.whispersystems.signalservice.internal.configuration.SignalServiceUrl
 import rx.Completable
 import rx.Observable
 import rx.Scheduler
 import rx.Single
-import rx.Subscription
 import rx.schedulers.Schedulers
 
 class SofaMessageManager(
+        private val userManager: UserManager,
         private val conversationStore: ConversationStore = ConversationStore(),
         private val baseApplication: BaseApplication = BaseApplication.get(),
-        private val trustStore: SignalTrustStore = SignalTrustStore(),
-        private val signalServiceUrl: SignalServiceUrl = SignalServiceUrl(baseApplication.getString(R.string.chat_url), trustStore),
-        private val signalServiceUrls: Array<SignalServiceUrl> = Array(1, { signalServiceUrl }),
-        private val protocolStore: ProtocolStore = ProtocolStore().init(),
-        private val signalPrefs: SignalPrefs = SignalPrefs,
-        private val userAgent: String = "Android " + BuildConfig.APPLICATION_ID + " - " + BuildConfig.VERSION_NAME + ":" + BuildConfig.VERSION_CODE,
+        private val chatManager: ChatManager = ChatManager(conversationStore, baseApplication, userManager),
         private val scheduler: Scheduler = Schedulers.io()
 ) {
 
-    private var wallet: HDWallet? = null
-    private var chatService: ChatService? = null
-    private var messageRegister: SofaMessageRegistration? = null
-    private var messageReceiver: SofaMessageReceiver? = null
-    private var messageSender: SofaMessageSender? = null
-    private var groupManager: SofaGroupManager? = null
-    private var connectivitySub: Subscription? = null
-
-    fun init(wallet: HDWallet): Completable {
-        this.wallet = wallet
-        return initEverything(wallet)
-    }
-
-    private fun initEverything(wallet: HDWallet): Completable {
-        initChatService(wallet)
-        initSenderAndReceiver(wallet)
-        return initRegistrationTask()
-                .doOnCompleted { attachConnectivityObserver() }
-    }
-
-    private fun initChatService(wallet: HDWallet) {
-        chatService = ChatService(signalServiceUrls, wallet, protocolStore, userAgent)
-    }
-
-    private fun initSenderAndReceiver(wallet: HDWallet) {
-        val messageSender = initMessageSender(wallet, protocolStore, conversationStore, signalServiceUrls)
-        this.messageReceiver = initMessageReceiver(wallet, protocolStore, conversationStore, signalServiceUrls, messageSender)
-        this.groupManager = initSofaGroupManager(messageSender, conversationStore)
-        this.messageSender = messageSender
-    }
-
-    private fun initMessageSender(wallet: HDWallet, protocolStore: ProtocolStore,
-                                  conversationStore: ConversationStore,
-                                  signalServiceUrls: Array<SignalServiceUrl>): SofaMessageSender {
-        return messageSender ?: SofaMessageSender(
-                wallet,
-                protocolStore,
-                conversationStore,
-                signalServiceUrls
-        )
-    }
-
-    private fun initMessageReceiver(wallet: HDWallet, protocolStore: ProtocolStore, conversationStore: ConversationStore,
-                                    signalServiceUrls: Array<SignalServiceUrl>, messageSender: SofaMessageSender): SofaMessageReceiver {
-        return messageReceiver ?: SofaMessageReceiver(
-                wallet,
-                protocolStore,
-                conversationStore,
-                signalServiceUrls,
-                messageSender
-        )
-    }
-
-    private fun initSofaGroupManager(messageSender: SofaMessageSender, conversationStore: ConversationStore): SofaGroupManager {
-        return SofaGroupManager(messageSender, conversationStore, baseApplication.userManager)
-    }
-
-    private fun attachConnectivityObserver() {
-        clearConnectivitySubscription()
-        connectivitySub =
-                baseApplication
-                .isConnectedSubject
-                .subscribeOn(scheduler)
-                .filter { isConnected -> isConnected }
-                .subscribe(
-                        { handleConnectivity() },
-                        { LogUtil.exception("Error checking connection state", it) }
-                )
-    }
-
-    private fun handleConnectivity() {
-        redoRegistrationTask()
-                .subscribeOn(scheduler)
-                .subscribe(
-                        { },
-                        { LogUtil.exception("Error during registration task", it) }
-                )
-    }
-
-    private fun initRegistrationTask(): Completable {
-        return if (messageRegister != null) return Completable.complete()
-        else initSofaMessageRegistration()
-    }
-
-    private fun initSofaMessageRegistration(): Completable {
-        val messageRegister = SofaMessageRegistration(chatService, protocolStore)
-        this.messageRegister = messageRegister
-        return messageRegister
-                .registerIfNeeded()
-                .doOnCompleted { messageReceiver?.receiveMessagesAsync() }
-    }
-
-    private fun redoRegistrationTask(): Completable {
-        val messageRegister = messageRegister ?: SofaMessageRegistration(chatService, protocolStore)
-        this.messageRegister = messageRegister
-        return messageRegister
-                .registerIfNeededWithOnboarding()
-                .doOnCompleted { messageReceiver?.receiveMessagesAsync() }
-    }
-
-    // Will send the message to a remote peer
-    // and store the message in the local database
-    fun sendAndSaveMessage(receiver: Recipient, message: SofaMessage) {
-        val messageTask = SofaMessageTask(receiver, message, SofaMessageTask.SEND_AND_SAVE)
-        messageSender?.addNewTask(messageTask)
-    }
-
-    // Will send the message to a remote peer
-    // but not store the message in the local database
-    fun sendMessage(recipient: Recipient, message: SofaMessage) {
-        val messageTask = SofaMessageTask(recipient, message, SofaMessageTask.SEND_ONLY)
-        messageSender?.addNewTask(messageTask)
-    }
-
-    // Will send an init message to remote peer
-    fun sendInitMessage(sender: User, recipient: Recipient) {
-        val initMessage = Init()
-                .setPaymentAddress(sender.paymentAddress)
-                .setLanguage(LocaleUtil.getLocale().language)
-        val messageBody = SofaAdapters.get().toJson(initMessage)
-        val sofaMessage = SofaMessage().makeNew(sender, messageBody)
-        val messageTask = SofaMessageTask(recipient, sofaMessage, SofaMessageTask.SEND_ONLY)
-        messageSender?.addNewTask(messageTask)
-    }
-
-    // Will store a transaction in the local database
-    // but not send the message to a remote peer. It will also save the state as "SENDING".
-    fun saveTransaction(user: User, message: SofaMessage) {
-        val recipient = Recipient(user)
-        val messageTask = SofaMessageTask(recipient, message, SofaMessageTask.SAVE_TRANSACTION)
-        messageSender?.addNewTask(messageTask)
-    }
-
-    // Updates a pre-existing message.
-    fun updateMessage(recipient: Recipient, message: SofaMessage) {
-        val messageTask = SofaMessageTask(recipient, message, SofaMessageTask.UPDATE_MESSAGE)
-        messageSender?.addNewTask(messageTask)
-    }
-
-    fun resendPendingMessage(sofaMessage: SofaMessage) = messageSender?.sendPendingMessage(sofaMessage)
-
-    fun createConversationFromGroup(group: Group): Single<Conversation> {
-        return groupManager
-                ?.createConversationFromGroup(group)
-                ?.subscribeOn(scheduler)
-                ?: Single.error(IllegalStateException("SofaGroupManager is null while createConversationFromGroup"))
-    }
-
-    fun updateConversationFromGroup(group: Group): Completable {
-        return groupManager
-                ?.updateConversationFromGroup(group)
-                ?.subscribeOn(scheduler)
-                ?: Completable.error(IllegalStateException("SofaGroupManager is null while updateConversationFromGroup"))
-    }
-
-    fun leaveGroup(group: Group): Completable {
-        return groupManager
-                ?.leaveGroup(group)
-                ?.subscribeOn(scheduler)
-                ?: Completable.error(IllegalStateException("SofaGroupManager is null while leaveGroup"))
-    }
-
-    fun resumeMessageReceiving() {
-        val registeredWithServer = signalPrefs.getRegisteredWithServer()
-        if (registeredWithServer && wallet != null) messageReceiver?.receiveMessagesAsync()
-    }
+    fun init(wallet: HDWallet): Completable = chatManager.initEverything(wallet)
 
     fun loadAllAcceptedConversations(): Single<List<Conversation>> {
         return conversationStore
@@ -345,7 +160,8 @@ class SofaMessageManager(
 
     fun rejectConversation(conversation: Conversation): Single<Conversation> {
         return if (conversation.isGroup) {
-            leaveGroup(conversation.recipient.group)
+            chatManager
+                    .leaveGroup(conversation.recipient.group)
                     .toSingle { conversation }
         } else baseApplication
                 .recipientManager
@@ -354,43 +170,49 @@ class SofaMessageManager(
                 .toSingle { conversation }
     }
 
-    fun tryUnregisterGcm(): Completable {
-        return messageRegister
-                ?.tryUnregisterGcm()
-                ?: Completable.error(IllegalStateException("Unable to register as class hasn't been initialised yet."))
+    fun leaveGroup(group: Group): Completable = chatManager.leaveGroup(group)
+
+    fun sendAndSaveMessage(receiver: Recipient, message: SofaMessage) {
+        chatManager.sendAndSaveMessage(receiver, message)
     }
 
-    fun forceRegisterChatGcm(): Completable {
-        return messageRegister
-                ?.registerChatGcm()
-                ?: Completable.error(IllegalStateException("Unable to register as class hasn't been initialised yet."))
+    fun sendMessage(recipient: Recipient, message: SofaMessage) {
+        chatManager.sendMessage(recipient, message)
     }
 
-    fun fetchLatestMessage(): Single<IncomingMessage> {
-        return messageReceiver
-                ?.fetchLatestMessage()
-                ?: Single.error(IllegalStateException("SofaMessageReceiver is null while fetchLatestMessage"))
+    fun sendInitMessage(sender: User, recipient: Recipient) {
+        chatManager.sendInitMessage(sender, recipient)
     }
 
-    fun clear() {
-        clearMessageReceiver()
-        clearMessageSender()
-        clearConnectivitySubscription()
+    fun saveTransaction(user: User, message: SofaMessage) {
+        chatManager.saveTransaction(user, message)
     }
 
-    private fun clearMessageReceiver() {
-        messageReceiver?.shutdown()
-        messageReceiver = null
+    fun updateMessage(recipient: Recipient, message: SofaMessage) {
+        chatManager.updateMessage(recipient, message)
     }
 
-    private fun clearMessageSender() {
-        messageSender?.clear()
-        messageSender = null
+    fun resendPendingMessage(sofaMessage: SofaMessage) = chatManager.resendPendingMessage(sofaMessage)
+
+    fun createConversationFromGroup(group: Group): Single<Conversation> {
+        return chatManager.createConversationFromGroup(group)
     }
 
-    private fun clearConnectivitySubscription() = connectivitySub?.unsubscribe()
+    fun updateConversationFromGroup(group: Group): Completable {
+        return chatManager.updateConversationFromGroup(group)
+    }
 
-    fun deleteSession() = protocolStore.deleteAllSessions()
+    fun resumeMessageReceiving() = chatManager.resumeMessageReceiving()
 
-    fun disconnect() = messageReceiver?.shutdown()
+    fun tryUnregisterGcm(): Completable = chatManager.tryUnregisterGcm()
+
+    fun forceRegisterChatGcm(): Completable = chatManager.forceRegisterChatGcm()
+
+    fun fetchLatestMessage(): Single<IncomingMessage> = chatManager.fetchLatestMessage()
+
+    fun clear() = chatManager.clear()
+
+    fun deleteSession() = chatManager.deleteSession()
+
+    fun disconnect() = chatManager.disconnect()
 }
